requirements.txt
numpy
xxhash
networkx
pygraphviz
matplotlib
pyyaml
orcapod @ git+https://github.com/walkerlab/orcapod@dev
